+++ src/main/java/br/com/zalf/prolog/webservice/integracao/IntegracaoDaoImpl.java
@@ -1,7 +1,8 @@
 package br.com.zalf.prolog.webservice.integracao;
 
 import br.com.zalf.prolog.webservice.commons.util.Log;
-import br.com.zalf.prolog.webservice.commons.util.StringUtils;
+import br.com.zalf.prolog.webservice.commons.util.PostgresUtils;
+import br.com.zalf.prolog.webservice.commons.util.SqlType;
 import br.com.zalf.prolog.webservice.database.DatabaseConnection;
 import br.com.zalf.prolog.webservice.integracao.agendador.os._model.IntegracaoConverter;
 import br.com.zalf.prolog.webservice.integracao.agendador.os._model.OsIntegracao;
@@ -12,7 +13,10 @@ import com.google.common.base.Preconditions;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.sql.*;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -429,32 +433,69 @@ public final class IntegracaoDaoImpl extends DatabaseConnection implements Integ
     }
 
     @Override
-    public void atualizaStatusOsIntegrada(@NotNull final Long codInternoOsProlog,
-                                          final boolean sucesso) throws Throwable {
-        atualizaStatusOsIntegrada(codInternoOsProlog, sucesso, null);
-    }
-
-    @Override
-    public void atualizaStatusOsIntegrada(@NotNull final Long codInternoOsProlog,
-                                          final boolean sucesso,
-                                          @Nullable final String errorMessage) throws Throwable {
+    public void atualizaStatusOsIntegrada(@NotNull final List<Long> codsInternoOsProlog,
+                                          final boolean pendente,
+                                          final boolean bloqueada,
+                                          final boolean incrementarTentativas) throws Throwable {
         Connection conn = null;
         PreparedStatement stmt = null;
         ResultSet rSet = null;
         try {
             conn = getConnection();
             stmt = conn.prepareStatement("select * from integracao.func_atualiza_status_os_integrada(" +
+                    "f_cods_interno_os_prolog => ?, " +
+                    "f_pendente => ?, " +
+                    "f_bloqueada => ?, " +
+                    "f_incrementar_tentativas => ?);");
+            stmt.setArray(1, PostgresUtils.listToArray(conn, SqlType.BIGINT, codsInternoOsProlog));
+            stmt.setBoolean(2, pendente);
+            stmt.setBoolean(3, bloqueada);
+            stmt.setBoolean(4, incrementarTentativas);
+            rSet = stmt.executeQuery();
+        } finally {
+            close(conn, stmt, rSet);
+        }
+    }
+
+    @Override
+    public void logarStatusOsComErro(@NotNull final Long codInternoOsProlog,
+                                     @Nullable final String errorMessage) throws Throwable {
+        Connection conn = null;
+        PreparedStatement stmt = null;
+        ResultSet rSet = null;
+        try {
+            conn = getConnection();
+            stmt = conn.prepareStatement("select * from integracao.func_atualiza_erro_os(" +
                     "f_cod_interno_os_prolog => ?, " +
-                    "f_sincronizado_sucesso => ?, " +
                     "f_error_message => ?);");
             stmt.setLong(1, codInternoOsProlog);
-            stmt.setBoolean(2, sucesso);
-            if (!sucesso) {
-                stmt.setString(3, StringUtils.trimToEmpty(errorMessage));
-            } else {
-                stmt.setNull(3, Types.VARCHAR);
-            }
+            stmt.setString(2, errorMessage);
             rSet = stmt.executeQuery();
+        } finally {
+            close(conn, stmt, rSet);
+        }
+    }
+
+    @Override
+    @NotNull
+    public List<Long> buscaCodOsByCodItem(@NotNull final List<Long> codItensProlog) throws Throwable {
+        Connection conn = null;
+        PreparedStatement stmt = null;
+        ResultSet rSet = null;
+        try {
+            conn = getConnection();
+            stmt = conn.prepareStatement("select * from integracao.func_busca_codigo_os(f_cod_itens_os => ?);");
+            stmt.setArray(1, PostgresUtils.listToArray(conn, SqlType.BIGINT, codItensProlog));
+            rSet = stmt.executeQuery();
+            if (rSet.next()) {
+                final List<Long> codsInternosOSsProlog = new ArrayList<>();
+                do {
+                    codsInternosOSsProlog.add(rSet.getLong("cod_interno_os_prolog"));
+                } while (rSet.next());
+                return codsInternosOSsProlog;
+            } else {
+                throw new Exception("Nenhum código de O.S encontrado para os itens: " + codItensProlog);
+            }
         } finally {
             close(conn, stmt, rSet);
         }