+++ src/main/java/br/com/zalf/prolog/webservice/integracao/agendador/os/IntegracaoOsTask.java
@@ -9,6 +9,7 @@ import br.com.zalf.prolog.webservice.integracao.avacorpavilan.requester.AvaCorpA
 import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 /**
@@ -17,6 +18,7 @@ import java.util.List;
  * @author Gustavo Navarro (https://github.com/gustavocnp95)
  */
 public class IntegracaoOsTask implements Runnable {
+
     @NotNull
     public static final String TAG = IntegracaoOsTask.class.getSimpleName();
     @NotNull
@@ -63,14 +65,17 @@ public class IntegracaoOsTask implements Runnable {
                         AvaCorpAvilanConverter.convert(osIntegracao));
                 Injection
                         .provideIntegracaoDao()
-                        .atualizaStatusOsIntegrada(osIntegracao.getCodInternoOsProlog(), true);
+                        .atualizaStatusOsIntegrada(
+                                Collections.singletonList(osIntegracao.getCodInternoOsProlog()),
+                                false,
+                                false,
+                                true);
             } catch (final Throwable t) {
                 // Não podemos fazer o throw nesse momento para não travar o fluxo de sincronia.
-                System.out.println("parar aqui");
                 try {
                     Injection
                             .provideIntegracaoDao()
-                            .atualizaStatusOsIntegrada(osIntegracao.getCodInternoOsProlog(), false, t.getMessage());
+                            .logarStatusOsComErro(osIntegracao.getCodInternoOsProlog(), t.getMessage());
                 } catch (final Throwable throwable) {
                     Log.e(TAG,
                             String.format("Erro ao atualizar o status da OS: %s", osIntegracao.getCodOsProlog()),
@@ -80,4 +85,5 @@ public class IntegracaoOsTask implements Runnable {
             }
         }
     }
+
 }